refactor: separate repository interfaces from their implementations
Move the concrete implementation classes ItemsRepositoryImpl and OrdersRepositoryImpl into their own dedicated files, while keeping only the interface definitions in ItemsRepository.kt and OrdersRepository.kt. This improves code organization, enhances separation of concerns, and increases modularity.
